64-bit fixes, after reports on the list. (Gilles Chehade/Thorsten Glaser/...)

+++ mip/regs.c
@@ -2141,7 +2141,7 @@ ngenregs(struct interpass *ipole)
 		nblock -= tempmin;
 		live = tmpalloc(BIT2BYTE(nbits));
 		RDEBUG(("nblock %p num %d size %zd\n",
-		    nblock, nbits, (int)(nbits * sizeof(REGW))));
+		    nblock, nbits, (size_t)(nbits * sizeof(REGW))));
 	}
